Hard-code in the proof of concept side data initialization bug.
If we do things this way then some ghost data is not actually set: it appears to
be corner points outside the physical domain. This is a bug with IBAMR since I
can see NaNs with the patches IBFEMethod uses for velocity interpolation - that
should not happen (i.e., we should always fill to the specified ghost width with
something reasonable).

The easiest way to see this is to run ex4 with the updated input file, break
on line 958 of lagrangian_interaction2d.f, and print u. For cell-centered data
all values are set correctly but for side-centered data there are NaNs (which I
set manually).